backend/celery-queue/requirements.txt
alembic==1.8.1
amqp==5.1.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.6
elastic-transport==8.4.0
elasticsearch==7.13.4
filelock==3.8.0
Flask==2.2.2
Flask-Migrate==2.5.2
flask-msearch==0.2.9.2
Flask-SQLAlchemy==2.5.1
flower==1.0.0
humanize==4.4.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.3
MarkupSafe==2.1.1
pipenv==2022.10.4
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==4.21.7
python-dotenv==0.21.0
pytz==2022.4
redis==3.4.1
requests==2.28.1
rq==1.2.2
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
tornado==6.2
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.2.2
Whoosh==2.7.4